Refactor - Extract CommonModels into separate file
