+++ moto/budgets/__init__.py
@@ -1,4 +1,5 @@
 from .models import budgets_backend
+from ..core.models import base_decorator
 
 budgets_backends = {"global": budgets_backend}
-mock_budgets = budgets_backend.decorator
+mock_budgets = base_decorator(budgets_backends)
