+++ moto/dynamodb_v20111205/__init__.py
@@ -1,4 +1,5 @@
 from .models import dynamodb_backend
+from ..core.models import base_decorator
 
 """
 An older API version of DynamoDB.
@@ -6,4 +7,4 @@ Please see the corresponding tests (tests/test_dynamodb_v20111205) on how to inv
 """
 
 dynamodb_backends = {"global": dynamodb_backend}
-mock_dynamodb = dynamodb_backend.decorator
+mock_dynamodb = base_decorator(dynamodb_backend)
