+++ moto/ec2/models/flow_logs.py
@@ -1,6 +1,6 @@
 import itertools
 from collections import defaultdict
-from moto.core.models import CloudFormationModel
+from moto.core import CloudFormationModel
 from ..exceptions import (
     FlowLogAlreadyExists,
     InvalidAggregationIntervalParameterError,
