+++ moto/glue/__init__.py
@@ -1,4 +1,5 @@
 from .models import glue_backend
+from ..core.models import base_decorator
 
 glue_backends = {"global": glue_backend}
-mock_glue = glue_backend.decorator
+mock_glue = base_decorator(glue_backends)
