+++ moto/iam/__init__.py
@@ -1,4 +1,5 @@
 from .models import iam_backend
+from ..core.models import base_decorator
 
 iam_backends = {"global": iam_backend}
-mock_iam = iam_backend.decorator
+mock_iam = base_decorator(iam_backends)
