+++ moto/route53/__init__.py
@@ -1,4 +1,5 @@
 from .models import route53_backend
+from ..core.models import base_decorator
 
 route53_backends = {"global": route53_backend}
-mock_route53 = route53_backend.decorator
+mock_route53 = base_decorator(route53_backends)
