+++ moto/s3/__init__.py
@@ -1,4 +1,5 @@
 from .models import s3_backend
+from ..core.models import base_decorator
 
 s3_backends = {"global": s3_backend}
-mock_s3 = s3_backend.decorator
+mock_s3 = base_decorator(s3_backends)
