+++ moto/ses/__init__.py
@@ -1,4 +1,5 @@
 from .models import ses_backend
+from ..core.models import base_decorator
 
 ses_backends = {"global": ses_backend}
-mock_ses = ses_backend.decorator
+mock_ses = base_decorator(ses_backends)
