+++ moto/sts/__init__.py
@@ -1,4 +1,5 @@
 from .models import sts_backend
+from ..core.models import base_decorator
 
 sts_backends = {"global": sts_backend}
-mock_sts = sts_backend.decorator
+mock_sts = base_decorator(sts_backends)
